+++ commons/dimas-com/src/builder/observable_builder.rs
@@ -94,6 +94,13 @@ where
 		self.feedback_interval = interval;
 		self
 	}
+
+	/// Set the session id.
+	#[must_use]
+	pub fn session_id(mut self, session_id: &str) -> Self {
+		self.session_id = session_id.into();
+		self
+	}
 }
 
 impl<P, CC, FC, EF, S> ObservableBuilder<P, NoSelector, CC, FC, EF, S>