+++ commons/dimas-com/src/lib.rs
@@ -13,8 +13,6 @@ pub mod communicator;
 pub mod enums;
 /// Modules errors
 pub mod error;
-/// a factory for creation of a communicator
-pub mod factory;
 /// `Communicator` trait
 pub mod traits;
 /// zenoh implementation
@@ -22,4 +20,3 @@ pub mod zenoh;
 
 // flatten
 pub use communicator::*;
-pub use factory::Factory;